refac(app): global machine provider
