+++ packages/app/src/systems/Account/hooks/index.tsx
@@ -1,2 +1,1 @@
 export * from "./useAccounts";
-export * from "./useIsLogged";
