+++ packages/app/src/systems/Core/components/PrivateRoute/PrivateRoute.tsx
@@ -1,7 +1,7 @@
 import type { ReactNode } from "react";
 import { Navigate } from "react-router-dom";
 
-import { useIsLogged } from "~/systems/Account";
+import { useIsLogged } from "../../hooks";
 
 type PrivateRouteProps = {
   children: ReactNode;
