+++ packages/app/src/systems/Core/components/Providers/index.tsx
@@ -1,2 +1,1 @@
 export * from "./Providers";
-export * from "./GlobalMachines";
