+++ packages/app/src/systems/Global/components/GlobalMachines/index.tsx
@@ -1,0 +1,1 @@
+export * from "./GlobalMachines";
