+++ packages/app/src/systems/Global/index.tsx
@@ -1,0 +1,1 @@
+export * from "./components";
